Split quota plan smoke fixtures
